split components and add static event (first draft)
