+++ family-day-out/src/components/EventItem.css
@@ -1,0 +1,18 @@
+.event-item {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    box-shadow: 0 2px 8px rgba(0,0,0,0.25);
+    padding: 0.5rem;
+    margin: 1rem 0;
+    border-radius: 12px;
+    margin-top: 50px;
+    /*background-color: #4b4b4b;*/
+}
+.event-item_description {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+    /*align-items: flex-end;*/
+    flex-flow:column;
+}
