+++ family-day-out/src/components/PlaceSearch.js
@@ -1,63 +1,8 @@
 import DateSelect from "./DateSelect";
 import Header from "./Header";
-import { useState } from "react";
-import PlacesAutocomplete, { geocodeByAddress, getLatLng } from "react-places-autocomplete"; 
-import Moment from 'moment';
+import PlacesAutocomplete from "react-places-autocomplete"; 
 
-const PlaceSearch = () => {
-  const [address, setAddress] = useState("");
-  const [startDate, setStartDate] = useState(new Date());
-  const [coordinates, setCoordinates] = useState({
-    lat: null,
-    lng: null
-  });
-
-  const handleSelect = async (value) => {
-    console.log(value);
-     const results = await geocodeByAddress(value);
-     const latLng = await getLatLng(results[0]);
-     setAddress(value);
-     setCoordinates(latLng);
-  };
-
-  //click on suggestion to show address
-  const handleClick = async (suggestion) => {
-    const results = await geocodeByAddress(suggestion.description);
-     const latLng = await getLatLng(results[0]);
-    const input = document.getElementById("id_address_autocomplete");
-    input.blur();
-    setAddress(suggestion.description);
-    setCoordinates(latLng);
-  };
-
-  const handleChange = (value) => {
-    setAddress(value);
-    setCoordinates({
-      lat: null,
-      lng: null
-    })
-  }
-//selected_days_weather_api
-const now = new Date();
-let Difference_In_Days = startDate.getTime() - now.getTime();
-let totalDaysInDays = Math.ceil(Difference_In_Days / (1000 * 3600 * 24));
-
-  const searchClick = () => {
-    if (coordinates.lat === null || coordinates.lng === null)
-    alert("Please select a place in the dropdown list");
-    fetch('http://0.0.0.0:5000/api/event_information', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          'crossDomain': true,
-        },
-        body: JSON.stringify({ 'place_address': address, 
-          'place_latitude': coordinates.lat, 
-          'place_longitude': coordinates.lng, 
-          'selected_date_event_api': Moment(startDate).format('DD-MM-YYYY'), 
-          'selected_days_weather_api': totalDaysInDays})
-        })
-      }
+const PlaceSearch = ({ address, handleClick, handleSelect, handleChange, searchClick, newStartDate, startDate }) => {
 
   return (
     <div><Header title="Family Day Out Planner" />
@@ -73,7 +18,7 @@ let totalDaysInDays = Math.ceil(Difference_In_Days / (1000 * 3600 * 24));
               {...getInputProps({ placeholder: "Search Places" })}
             />
             <div className="react-datepicker-container">
-            <DateSelect newStartDate={setStartDate} startDate={startDate} />
+            <DateSelect newStartDate={newStartDate} startDate={startDate} />
             </div>
             <button onClick={searchClick} className="btn">Search</button>
             <div className="autocomplete-dropdown-container">
